Probar dotDensity 'single' para el ticket adhesivo vía QZ Tray
Con 'double' (igual que el recibo térmico) el ticket seguía saliendo
igual de escalonado que por el diálogo de Windows, ya confirmado con
la impresora real conectada — indicio de que este modelo de impresora
de etiquetas no maneja bien el modo de doble densidad.

+++ app-sastreria/src/app/core/services/qz-print.service.ts
@@ -227,7 +227,12 @@ export class QzPrintService {
         type: 'raw',
         format: 'image',
         data: dataUrl,
-        options: { language: 'ESCPOS', dotDensity: 'double', quantization: 'black', threshold: 128 },
+        // dotDensity 'single' (a diferencia de 'double' en el recibo térmico):
+        // esta impresora de etiquetas es un equipo físico distinto al de
+        // recibos, y con 'double' salía igual de escalonada que por el
+        // diálogo de impresión de Windows — indicio de que este modelo no
+        // maneja bien el modo de doble densidad del comando ESC/POS.
+        options: { language: 'ESCPOS', dotDensity: 'single', quantization: 'black', threshold: 128 },
       },
     ]);
   }
